% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_data_up_to_date.R
\name{assert_data_up_to_date}
\alias{assert_data_up_to_date}
\title{Checks latency of input data.}
\usage{
assert_data_up_to_date(
  data,
  expected_max_time_value,
  target_label = checkmate::vname(data),
  location_col_name = "geo_value",
  date_col_name = "time_value",
  overlatent_err_thresh = 0.2
)
}
\arguments{
\item{data}{Data frame containing target data.}

\item{expected_max_time_value}{Date. The most recent date for which
observations are expected.}

\item{target_label}{Character. Human-readable label for the target.
Default is the variable name of the data object.}

\item{location_col_name}{Character. Name of the location column.
Default is "geo_value".}

\item{date_col_name}{Character. Name of the date column.
Default is "time_value".}

\item{overlatent_err_thresh}{Numeric. Proportion threshold for raising error
vs warning. Default 0.20.}
}
\value{
Invisible NULL. Raises warnings or errors
based on the proportion of locations with excess latency.
}
\description{
Checks latency of input data.
}
