% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_authorized_users.R
\name{check_authorized_users}
\alias{check_authorized_users}
\title{Check if a user is authorized to modify model
directories. This function verifies whether a GitHub
user is authorized to modify specific model directories
in a Hub by checking the designated users in model
metadata.}
\usage{
check_authorized_users(changed_dirs, gh_actor, base_hub_path)
}
\arguments{
\item{changed_dirs}{Character vector. Directory names
that have been changed.}

\item{gh_actor}{Character. GitHub username of the person
making changes.}

\item{base_hub_path}{Character. Path to the base hub
directory.}
}
\value{
Exits invisibly, with error if the user is
unauthorized and prints success message if authorized.
}
\description{
Check if a user is authorized to modify model
directories. This function verifies whether a GitHub
user is authorized to modify specific model directories
in a Hub by checking the designated users in model
metadata.
}
