% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hub_baselines.R
\name{check_data_latency}
\alias{check_data_latency}
\title{Checks data latency in the input epi_df}
\usage{
check_data_latency(
  epi_df,
  reference_date,
  desired_max_time_value,
  target_label,
  overlatent_err_thresh = 0.2
)
}
\arguments{
\item{epi_df}{An epi_df object of time series data.}

\item{reference_date}{Date. The reference date for the forecast.}

\item{desired_max_time_value}{Date. The most recent date for which
observations are expected.}

\item{target_label}{Character. Human-readable label for the target,
e.g., "Hospital Admissions".}

\item{overlatent_err_thresh}{Numeric. Proportion threshold of locations
with excess latency to raise an error. Default 0.20 (20\%).}
}
\value{
NULL. Raises a warning if any location has excess latency or
raises an error if proportion of locations with excess latency exceeds threshold.
}
\description{
Checks data latency in the input epi_df
}
