% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{default_excluded_locations}
\alias{default_excluded_locations}
\title{Default US state/territory abbreviations excluded
from hub data.}
\format{
An object of class \code{character} of length 5.
}
\usage{
default_excluded_locations
}
\description{
Excludes Virgin Islands (VI), Guam (GU),
American Samoa (AS), Northern Mariana Islands (MP),
and Minor Outlying Islands (UM).
}
\keyword{datasets}
