% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{excluded_locations}
\alias{excluded_locations}
\title{Two digits FIPS codes for locations excluded from Hubs'
target data.}
\format{
An object of class \code{character} of length 5.
}
\usage{
excluded_locations
}
\description{
Excludes Virgin Islands (78), Northern Mariana
Islands (69), Guam (66), American Samoa (60), and Minor
Outlying Islands (74).
}
\keyword{datasets}
