% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_oracle_output.R
\name{flatten_task}
\alias{flatten_task}
\title{Transform a modeling task represented as a nested list
to a single data frame}
\usage{
flatten_task(task)
}
\arguments{
\item{task}{Nested list representing a modeling task,
as one entry of the output of \code{\link[hubUtils:get_round_model_tasks]{hubUtils::get_round_model_tasks()}}.
Must have a \code{target_end_date} specification.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} of all potentially
valid submittable outputs for the modeling task defined in \code{task}.
Each row of the table represents a single valid forecastable quantity
(e.g. "\code{target} X on \code{target_end_date} Y in \code{location} Z"),
plus a valid submittable output_type for forecasting that quantity.
If multiple \code{output_type}s are accepted for a given valid forecastable
quantity, that quantity will be represented multiple times, with
one row for each valid associated \code{output_type}.
}
\description{
Transform a modeling task represented as a nested list
to a single data frame
}
