% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_oracle_output.R
\name{flatten_task_list}
\alias{flatten_task_list}
\title{Transform a group of modeling task represented as a list of
nested lists into a single data frame.}
\usage{
flatten_task_list(task_list, .deduplicate = TRUE)
}
\arguments{
\item{task_list}{List of tasks. Each entry should itself be
be a nested list that can be passed to \code{\link[=flatten_task]{flatten_task()}}.}

\item{.deduplicate}{deduplicate the output if the same flat
configuration is found multiple times while flattening the task list?
Default \code{TRUE}.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} of all potentially
valid submittable outputs for all the modeling tasks defined in \code{task_lists}.
Each row of the table represents a single valid forecastable quantity
(e.g. "\code{target} X on \code{target_end_date} Y in \code{location} Z"),
plus a valid submittable output_type for forecasting that quantity.
If multiple \code{output_type}s are accepted for a given valid forecastable
quantity, that quantity will be represented multiple times, with
one row for each valid associated \code{output_type}.
}
\description{
Calls \code{\link[=flatten_task]{flatten_task()}} on each entry of the task list.
}
