% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_oracle_output.R
\name{generate_oracle_output}
\alias{generate_oracle_output}
\title{Generate and save oracle output for the Hub}
\usage{
generate_oracle_output(hub_path)
}
\arguments{
\item{hub_path}{Path to the hub root.}
}
\value{
nothing, invisibly, on success.
}
\description{
Generate and save oracle output for the Hub
}
