% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_alternate_target_name}
\alias{get_alternate_target_name}
\title{Get alternate target name from full target string.}
\usage{
get_alternate_target_name(target)
}
\arguments{
\item{target}{Character. Full target name (e.g.,
"wk inc covid hosp").}
}
\value{
Character. Short target name ("hosp", "prop_ed",
or NA if unrecognized).
}
\description{
Converts full target strings to short names for use in
data outputs.
}
