% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_disease_name}
\alias{get_disease_name}
\title{Get display name for a given disease.}
\usage{
get_disease_name(disease)
}
\arguments{
\item{disease}{Character. Disease identifier ("covid",
"rsv", or "flu").}
}
\value{
Character. Display name (e.g., "COVID-19", "RSV",
"Influenza").
}
\description{
Converts disease identifier to human-readable display
name.
}
