% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hub_repo}
\alias{get_hub_repo}
\title{Get hub repository name for a given disease.}
\usage{
get_hub_repo(disease)
}
\arguments{
\item{disease}{Character. Disease identifier ("covid" or "rsv").}
}
\value{
Character. Repository name.
}
\description{
Converts disease identifier to GitHub repository name.
}
