% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hub_repo_name}
\alias{get_hub_repo_name}
\title{Get GitHub repository name for a given disease.}
\usage{
get_hub_repo_name(disease)
}
\arguments{
\item{disease}{Character. Disease identifier ("covid"
or "rsv").}
}
\value{
Character. GitHub repository name.
}
\description{
Converts disease identifier to corresponding GitHub
repository name.
}
