% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hub_repo_url}
\alias{get_hub_repo_url}
\title{Get GitHub repository URL for a given disease.}
\usage{
get_hub_repo_url(disease)
}
\arguments{
\item{disease}{Character. Disease identifier ("covid",
"rsv", or "flu").}
}
\value{
Character. Full HTTPS GitHub URL.
}
\description{
Constructs full HTTPS GitHub URL for the forecast
hub corresponding to the given disease identifier.
}
