% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nhsn_col_name}
\alias{get_nhsn_col_name}
\title{Get NHSN column name for a given disease}
\usage{
get_nhsn_col_name(disease)
}
\arguments{
\item{disease}{Disease name ("covid", "rsv", or "flu")}
}
\value{
Character string with the NHSN column name
}
\description{
Get NHSN column name for a given disease
}
