% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nssp_col_name}
\alias{get_nssp_col_name}
\title{Get NSSP column name for a given disease}
\usage{
get_nssp_col_name(disease)
}
\arguments{
\item{disease}{Disease name ("covid", "rsv", or "flu")}
}
\value{
Character string with the NSSP column name
}
\description{
Get NSSP column name for a given disease
}
