% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_reporting_minimum.R
\name{get_reference_date_ens_minimum}
\alias{get_reference_date_ens_minimum}
\title{Get a hub's ensemble reporting minimum designated
models for one reference date.}
\usage{
get_reference_date_ens_minimum(hub_reports_path, disease, reference_date)
}
\arguments{
\item{hub_reports_path}{character, path to the
forecast hub reports directory.}

\item{disease}{character, disease name ("covid" or
"rsv").}

\item{reference_date}{character or Date, the
reference date to look up.}
}
\value{
The minimum, as an integer.
}
\description{
Get a hub's ensemble reporting minimum designated
models for one reference date.
}
