% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{get_representative_round_ids}
\alias{get_representative_round_ids}
\title{Get a vector of representative round ids for a modeling hub}
\usage{
get_representative_round_ids(config_tasks)
}
\arguments{
\item{config_tasks}{List of Hub tasks, as the
output of \code{\link[hubUtils:read_config]{hubUtils::read_config()}} or the \code{config_tasks}
attribute of a \code{\link[hubData:connect_hub]{<hub_connection>}}.}
}
\value{
Representative round ids, as a vector.
}
\description{
Wrapper of \code{\link[hubUtils:get_round_idx]{hubUtils::get_round_ids()}} that returns a
flat vector of round_id values that fully specify and uniquely index
the set of available task ids regardless of whether \code{round_id_from_variable: true}.
}
\details{
Functions such as \code{\link[hubValidations:expand_model_out_grid]{hubValidations::expand_model_out_grid()}} will create
duplicate values when \code{round_id_from_variable: true} if they are called on all
valid round_id values, since what we actually want to do is call them once
for each unique \code{round_idx}. But since \code{\link[hubValidations:expand_model_out_grid]{hubValidations::expand_model_out_grid()}}
expects a round_id rather than a round_idx for input, the solution is to
compute a set of \emph{representative} \verb{round id} values, one for each unique \code{round_idx}.
Note that when \code{round_id_from_variable: false}, this should simply by the set of
\code{round_id} values.
}
\examples{

hub_con <- hubData::connect_hub(
 system.file("testhubs/simple", package = "hubUtils")
)
config_tasks <- attr(hub_con, "config_tasks")

# multiple round ids derived from `reference_date` values via
# `round_id_from_variable: true`
all_round_ids <- hubUtils::get_round_ids(config_tasks)
all_round_ids

# But these all correspond to a single `round_idx`:
purrr::map_vec(all_round_ids, \(id) hubUtils::get_round_idx(config_tasks, id))

# get representative round ids returns a single "representative" id
# corresponding to this single `round_idx`:
get_representative_round_ids(config_tasks)

}
