% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{get_round_ids_vec}
\alias{get_round_ids_vec}
\title{Get a vector of unique round ids for a modeling hub}
\usage{
get_round_ids_vec(config_tasks)
}
\arguments{
\item{config_tasks}{List of Hub tasks, as the
output of \code{\link[hubUtils:read_config]{hubUtils::read_config()}} or the \code{config_tasks}
attribute of a \code{\link[hubData:connect_hub]{<hub_connection>}}.}
}
\value{
Unique round ids, as a vector.
}
\description{
Wrapper of \code{\link[hubUtils:get_round_ids]{hubUtils::get_round_ids()}} that returns a
flat vector of unique round_id values regardless of whether
\code{round_id_from_variable} is \code{TRUE} or \code{FALSE}.
}
\details{
\code{\link[hubUtils:get_round_ids]{hubUtils::get_round_ids()}} with \code{flatten = "all"} returns a vector,
but may duplicate round_ids when \code{round_id_from_variable: true}.
\code{\link[hubUtils:get_round_ids]{hubUtils::get_round_ids()}} with \code{flatten = "model_task"} handles
both \code{round_id_from_variable: true} and \code{round_id_from_variable: false},
but returns a list. This function provides a workaround by calling
\code{\link[purrr:list_c]{purrr::list_c()}} on the output of \code{\link[hubUtils:get_round_ids]{hubUtils::get_round_ids()}} with
\code{flatten = "model_task"}.
}
\examples{
hub_con <- hubData::connect_hub(
 system.file("testhubs/simple", package = "hubUtils")
)
config_tasks <- attr(hub_con, "config_tasks")

get_round_ids_vec(config_tasks)

}
