% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_target_data_type}
\alias{get_target_data_type}
\title{Get target data type from full target string.}
\usage{
get_target_data_type(target)
}
\arguments{
\item{target}{Character. Full target name (e.g.,
"wk inc covid hosp").}
}
\value{
Character. Target data type ("hosp", "prop_ed",
or NA if unrecognized).
}
\description{
Converts full target strings to data type for use in
data outputs.
}
