% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_hub_targets}
\alias{get_unique_hub_targets}
\title{Get unique targets from hub time-series data.}
\usage{
get_unique_hub_targets(base_hub_path)
}
\arguments{
\item{base_hub_path}{Path to the base hub directory.}
}
\value{
Character vector of unique target names.
}
\description{
Reads the target-data time-series and returns the
unique target names present in the data.
}
