% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{included_locations}
\alias{included_locations}
\title{Two digits FIPS codes for locations included in Hubs'
target data.}
\format{
An object of class \code{character} of length 53.
}
\usage{
included_locations
}
\description{
Includes US states, DC, and Puerto Rico (PR). Excludes
Virgin Islands (78), Northern Mariana Islands (69),
Guam (66), American Samoa (60), and Minor Outlying
Islands (74).
}
\keyword{datasets}
