% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_ed_target}
\alias{is_ed_target}
\title{Check if target is an emergency department visits proportion.}
\usage{
is_ed_target(target)
}
\arguments{
\item{target}{Character. Target name to check.}
}
\value{
Logical. TRUE if target ends with "prop ed visits", FALSE otherwise.
}
\description{
Helper function to identify targets that end with "prop ed visits".
}
