% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_reporting_minimum.R
\name{n_models_for_ens_reporting_as_of}
\alias{n_models_for_ens_reporting_as_of}
\title{Look up the minimum designated submissions required
to report an ensemble for a reference date.}
\usage{
n_models_for_ens_reporting_as_of(reference_date)
}
\arguments{
\item{reference_date}{character or Date, the
reference date to look up.}
}
\value{
The minimum, as an integer.
}
\description{
Look up the minimum designated submissions required
to report an ensemble for a reference date.
}
