% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_data}
\alias{population_data}
\title{Population Data by Location}
\format{
A data frame with 2 variables:
\describe{
\item{location}{Location identifier, two digits FIPS codes}
\item{population}{Population count from 2023 census}
}
}
\source{
data-raw/population_data.R
}
\usage{
population_data
}
\description{
A dataset containing population information for US states and national level based on
2023 census data.
}
\keyword{datasets}
