% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{recognized_diseases}
\alias{recognized_diseases}
\title{Recognized disease identifiers.}
\format{
An object of class \code{character} of length 3.
}
\usage{
recognized_diseases
}
\description{
Character vector of all disease identifiers recognized
by hubhelpr utility functions. Includes diseases without
full forecast pipeline support.
}
\keyword{datasets}
