% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{supported_diseases}
\alias{supported_diseases}
\title{Supported disease identifiers.}
\format{
An object of class \code{character} of length 2.
}
\usage{
supported_diseases
}
\description{
Character vector of disease identifiers with full
forecast pipeline support (baseline and ensemble
generation).
}
\keyword{datasets}
