% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_authorized_users.R
\name{update_authorized_users}
\alias{update_authorized_users}
\title{Get authorized GitHub users from Hub model-metadata.}
\usage{
update_authorized_users(base_hub_path)
}
\arguments{
\item{base_hub_path}{Path to the base hub directory.}
}
\value{
\code{NULL}, invisibly. Writes
\code{authorized_users.json} file to the \code{auxiliary-data}
directory of the hub as a side effect.
}
\description{
This function reads model metadata from a Hub and extracts
authorized GitHub users for each model, saving the results
to a JSON file in the hub's \code{auxiliary-data} directory.
Models without authorized users will have an empty array
in the output.
}
