% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_oracle_output.R
\name{write_oracle_output}
\alias{write_oracle_output}
\alias{generate_oracle_output}
\title{Generate and save oracle output for the Hub}
\usage{
write_oracle_output(
  hub_path,
  output_dirpath = fs::path(hub_path, "target-data"),
  ts_date_col = "date"
)

generate_oracle_output(
  hub_path,
  output_dirpath = fs::path(hub_path, "target-data"),
  ts_date_col = "date"
)
}
\arguments{
\item{hub_path}{Path to the hub root.}

\item{output_dirpath}{Directory in which to
write the oracle-output.parquet file. Defaults
to the \code{target-data} subdirectory of the hub specified
in \code{hub_path}.}

\item{ts_date_col}{Name of the date column in
the \code{time-series} format target data. Will be converted
to \code{target_end_date} for the oracle output. Default \code{date}.}
}
\value{
nothing, invisibly, on success.
}
\description{
Generate and save oracle output for the Hub
}
