% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_authorized_users.R
\name{check_authorized_users}
\alias{check_authorized_users}
\title{Check if a user is authorized to modify model
directories.}
\usage{
check_authorized_users(changed_model_ids, gh_actor, base_hub_path)
}
\arguments{
\item{changed_model_ids}{Character vector. Model IDs that
have been modified.}

\item{gh_actor}{Character. GitHub username of the person
making changes.}

\item{base_hub_path}{Character. Path to the base hub
directory.}
}
\value{
\code{NULL}, invisibly, raising an error if the user is
unauthorized and printing a success message if the user is authorized.
}
\description{
This function verifies whether a GitHub
user is authorized to modify specific model IDs
in a Hub by checking the designated users in model
metadata.
}
