% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_latency.R
\name{check_data_latency}
\alias{check_data_latency}
\title{Checks latency of input data.}
\usage{
check_data_latency(
  data,
  target_label,
  expected_max_time_value,
  location_col_name = "geo_value",
  date_col_name = "time_value",
  overlatent_err_thresh = 0.2
)
}
\arguments{
\item{data}{Data frame containing target data.}

\item{target_label}{Character. Human-readable label for the target.}

\item{expected_max_time_value}{Date. The most recent date for which
observations are expected.}

\item{location_col_name}{Character. Name of the location column.
Default is "geo_value".}

\item{date_col_name}{Character. Name of the date column.
Default is "time_value".}

\item{overlatent_err_thresh}{Numeric. Proportion threshold for raising error
vs warning. Default 0.20.}
}
\value{
Invisible NULL. Raises warnings or errors
based on the proportion of locations with excess latency.
}
\description{
Checks latency of input data.
}
