% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{excluded_locations}
\alias{excluded_locations}
\title{Two digits FIPS codes for locations excluded from Hubs'
target data.}
\format{
An object of class \code{character} of length 5.
}
\usage{
excluded_locations
}
\description{
Two digits FIPS codes for locations excluded from Hubs'
target data.
}
\keyword{datasets}
