% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hub_baselines.R
\name{generate_hub_baseline}
\alias{generate_hub_baseline}
\title{Generate hub baseline forecasts for a given disease and reference date}
\usage{
generate_hub_baseline(base_hub_path, reference_date, disease)
}
\arguments{
\item{base_hub_path}{Path to the base hub directory.}

\item{reference_date}{Reference date (should be a Saturday).}

\item{disease}{Disease name ("covid" or "rsv").}
}
\description{
Generate hub baseline forecasts for a given disease and reference date
}
