% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hub_name}
\alias{get_hub_name}
\title{Get hub name for a given disease.}
\usage{
get_hub_name(disease)
}
\arguments{
\item{disease}{Character. Disease identifier ("covid"
or "rsv").}
}
\value{
Character. Hub name (e.g., "CovidHub", "RSVHub").
}
\description{
Converts disease identifier to proper hub name format.
}
