% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hub_repo}
\alias{get_hub_repo}
\title{Get GitHub repository path for a given disease.}
\usage{
get_hub_repo(disease)
}
\arguments{
\item{disease}{Character. Disease identifier ("covid" or "rsv").}
}
\value{
Character. GitHub repository path.
}
\description{
Converts disease identifier to corresponding GitHub repository path.
}
