% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hub_target_data.R
\name{get_hubverse_format_nhsn_data}
\alias{get_hubverse_format_nhsn_data}
\title{Get and format NHSN data for a given disease.}
\usage{
get_hubverse_format_nhsn_data(
  disease,
  as_of = lubridate::today(),
  start_date = NULL,
  end_date = NULL,
  date_col_name = "target_end_date"
)
}
\arguments{
\item{disease}{Disease name ("covid" or "rsv").}

\item{as_of}{As-of date of the data pull. Default is
the system date as determined by \code{\link[lubridate:today]{lubridate::today()}}.}

\item{start_date}{First week-ending
date to include for the NHSN dataset. Default value
is NULL (no filtering).}

\item{end_date}{Last week-ending
date to include for the NHSN dataset. Default value
is NULL (no filtering).}

\item{date_col_name}{Character. Name for the date column
in the output. Default is "target_end_date".}
}
\value{
Data frame with formatted NHSN data.
}
\description{
This function pulls the NHSN hospital admissions
data, formats and returns it in the hubverse format.
Missing observations before each location's first
reported week within the requested date range are
dropped.
}
