% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hub_target_data.R
\name{get_hubverse_format_nhsn_data}
\alias{get_hubverse_format_nhsn_data}
\title{Get and format NHSN data for a given disease.}
\usage{
get_hubverse_format_nhsn_data(
  disease,
  as_of = lubridate::today(),
  included_locations = hubhelpr::included_locations,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{disease}{Disease name ("covid" or "rsv").}

\item{as_of}{As-of date of the data pull. Default is
the system date as determined by \code{\link[lubridate:now]{lubridate::today()}}.}

\item{included_locations}{Vector of location codes to
include in the output.
Default value \code{hubhelpr::included_locations}.}

\item{start_date}{First week-ending
date to include for the NHSN dataset. Default value
is NULL (no filtering).}

\item{end_date}{Last week-ending
date to include for the NHSN dataset. Default value
is NULL (no filtering).}
}
\value{
Data frame with formatted NHSN data.
}
\description{
This function pulls the NHSN hospital admissions data,
formats and returns it in the hubverse format.
}
