% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hub_target_data.R
\name{get_hubverse_format_nssp_data}
\alias{get_hubverse_format_nssp_data}
\title{Get and format NSSP data for a given disease.}
\usage{
get_hubverse_format_nssp_data(
  disease,
  base_hub_path,
  as_of = lubridate::today(),
  included_locations = hubhelpr::included_locations,
  nssp_update_local = FALSE,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{disease}{Disease name ("covid" or "rsv").}

\item{base_hub_path}{Path to the base hub directory.}

\item{as_of}{As-of date of the data pull. Default is
the system date as determined by \code{\link[lubridate:now]{lubridate::today()}}.}

\item{included_locations}{Vector of location codes to
include in the output.
Default value \code{hubhelpr::included_locations}.}

\item{nssp_update_local}{Logical. Whether to update NSSP
data from local file \code{auxiliary-data/latest.parquet}
(default: FALSE).}

\item{start_date}{First week-ending
date to include for the NSSP dataset. Default value
is NULL (no filtering).}

\item{end_date}{Last week-ending
date to include for the NSSP dataset. Default value
is NULL (no filtering).}
}
\value{
Data frame with formatted NSSP data.
}
\description{
This function pulls, formats and returns NSSP emergency
department visit data in hubverse format.
}
