% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_targets}
\alias{get_unique_targets}
\title{Get unique targets from hub time-series data.}
\usage{
get_unique_targets(base_hub_path, disease)
}
\arguments{
\item{base_hub_path}{Path to the base hub directory.}

\item{disease}{Character. Disease identifier ("covid" or
"rsv").}
}
\value{
Character vector of unique full target names.
}
\description{
Reads the target-data time-series and extracts unique
full target names (e.g., "wk inc covid hosp",
"wk inc covid prop ed visits") for a given disease.
}
