% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubhelpr-package.R
\docType{package}
\name{hubhelpr-package}
\alias{hubhelpr}
\alias{hubhelpr-package}
\title{hubhelpr: Helper functions for CDC hubs maintenance}
\description{
This package provides functions to help with the maintenance of CDC hubs.
}
\author{
\strong{Maintainer}: Subekshya Bidari \email{zib2@cdc.gov}

Authors:
\itemize{
  \item Dylan Morris \email{dzl1@cdc.gov}
  \item Trevor Martin \email{upx3@cdc.gov}
}

}
\keyword{internal}
