% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{supported_viz_datasets}
\alias{supported_viz_datasets}
\title{Supported data sources for visualization target data.}
\format{
An object of class \code{character} of length 2.
}
\usage{
supported_viz_datasets
}
\description{
Character vector of valid dataset identifiers that can
be passed to \code{\link{write_viz_target_data}}.
}
\keyword{datasets}
