% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hub_target_data.R
\name{update_hub_target_data}
\alias{update_hub_target_data}
\title{Get and process target data for the hub}
\usage{
update_hub_target_data(
  base_hub_path,
  disease,
  as_of = lubridate::today(),
  nhsn_first_weekending_date = lubridate::as_date("2024-11-09"),
  excluded_locations = hubhelpr::excluded_locations,
  legacy_file = FALSE
)
}
\arguments{
\item{base_hub_path}{Path to the base hub directory.}

\item{disease}{Disease name ("covid" or "rsv").}

\item{as_of}{As-of date of the data pull. Default is the system date as determined by \code{\link[lubridate:now]{lubridate::today()}}.}

\item{nhsn_first_weekending_date}{First week-ending date to include for
the NHSN dataset. Default value is "2024-11-09".}

\item{excluded_locations}{Vector of location codes to exclude from the output.
Default value \code{hubhelpr::excluded_locations}.}

\item{legacy_file}{Logical. Whether to write legacy CSV output (default: FALSE).}
}
\value{
Writes \code{time-series.parquet} and optionally legacy csv target data files
to the target-data directory in the hub.
}
\description{
This function pulls, formats and save NHSN and NSSP data for use in the hub.
}
