% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_viz_target_data.R
\name{write_viz_target_data}
\alias{write_viz_target_data}
\title{Write target data for visualization to disk.}
\usage{
write_viz_target_data(
  reference_date,
  base_hub_path,
  hub_reports_path,
  disease,
  target,
  included_locations = hubhelpr::included_locations
)
}
\arguments{
\item{reference_date}{Character, the reference date for
the forecast in YYYY-MM-DD format (ISO-8601).}

\item{base_hub_path}{Character, path to the forecast hub
directory.}

\item{hub_reports_path}{Character, path to forecast hub
reports directory.}

\item{disease}{Character, disease name ("covid" or "rsv").}

\item{target}{Character, target name to filter
(e.g., "wk inc covid hosp").}

\item{included_locations}{Character vector of location
codes to include in the output. Default
hubhelpr::included_locations.}
}
\value{
Invisibly returns file path where data was
written.
}
\description{
This function reads target data from the hub's
time-series file, filters to the latest vintage and
specified target, and writes it to the weekly-summaries
directory for use by the visualization webpage.
}
